Additional cleanups. Introduced a base class for canonical and relative references to avoid code duplication.
