+++ ge/org.osate.ge/src/org/osate/ge/CanonicalBusinessObjectReference.java
@@ -23,14 +23,8 @@
  */
 package org.osate.ge;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
-
 import org.osate.ge.businessobjecthandling.BusinessObjectHandler;
 import org.osate.ge.internal.services.impl.ReferenceEncoder;
-
-import com.google.common.collect.ImmutableList;
 
 /**
  * Immutable data type for canonical references to a business object. Canonical business object references are created for a
@@ -40,69 +34,14 @@ import com.google.common.collect.ImmutableList;
  * @since 2.0
  *
  */
-public final class CanonicalBusinessObjectReference {
-	private ImmutableList<String> segments;
-	private ImmutableList<String> lcSegments; // Lowercase segments. Used for comparison.
-
+public final class CanonicalBusinessObjectReference extends BusinessObjectReference {
 	/**
 	 * Creates an instance from an array of segments. Segments are case insensitive. Throws an exception is optional or if the segments
 	 * array is null or empty.
 	 * @param segments is an array of segments that makes up the reference.
 	 */
 	public CanonicalBusinessObjectReference(final String... segments) {
-		if (segments == null || segments.length < 1) {
-			throw new IllegalArgumentException("segments must contain at least one segment");
-		}
-
-		// Check that all segments are non-null
-		for (final String seg : segments) {
-			if (seg == null) {
-				throw new IllegalArgumentException("segment is null");
-			}
-		}
-
-		// Store segments and lower case segments in separate lists. The lowercase segments are used for comparison while the other list is used to preserve
-		// case for serialization.
-		this.segments = ImmutableList.copyOf(segments);
-		this.lcSegments = this.segments.stream().map(s -> s.toLowerCase(Locale.ROOT))
-				.collect(ImmutableList.toImmutableList());
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((lcSegments == null) ? 0 : lcSegments.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj) {
-			return true;
-		}
-		if (obj == null) {
-			return false;
-		}
-		if (getClass() != obj.getClass()) {
-			return false;
-		}
-
-		final CanonicalBusinessObjectReference other = (CanonicalBusinessObjectReference)obj;
-		return lcSegments.equals(other.lcSegments);
-	}
-
-	@Override
-	public String toString() {
-		return Arrays.toString(segments.toArray());
-	}
-
-	/**
-	 * Returns an immutable list containing the segments.
-	 * @return an immutable list of segments.
-	 */
-	public List<String> getSegments() {
-		return segments;
+		super(segments);
 	}
 
 	/**
@@ -112,16 +51,16 @@ public final class CanonicalBusinessObjectReference {
 	 * @return the encoded reference
 	 */
 	public String encode() {
-		return ReferenceEncoder.encode(segments);
+		return ReferenceEncoder.encode(getSegments());
 	}
 
 	/**
 	 * Converts the instance to an instance of the serialized diagram model type.
 	 * @return the converted object.
 	 */
-	public org.osate.ge.diagram.CanonicalBusinessObjectReference toMetamodel() {
+	public final org.osate.ge.diagram.CanonicalBusinessObjectReference toMetamodel() {
 		final org.osate.ge.diagram.CanonicalBusinessObjectReference newValue = new org.osate.ge.diagram.CanonicalBusinessObjectReference();
-		for(final String seg : segments) {
+		for (final String seg : getSegments()) {
 			newValue.getSeg().add(seg);
 		}
 		return newValue;
