+++ ge/org.osate.ge/src/org/osate/ge/RelativeBusinessObjectReference.java
@@ -23,13 +23,7 @@
  */
 package org.osate.ge;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
-
 import org.osate.ge.businessobjecthandling.BusinessObjectHandler;
-
-import com.google.common.collect.ImmutableList;
 
 /**
  * Immutable data type for relative references to a business object. Relative business object references are created for a
@@ -37,95 +31,23 @@ import com.google.common.collect.ImmutableList;
  * A relative reference along with a containing diagram element must uniquely identify the business object.
  * @since 2.0
  */
-public class RelativeBusinessObjectReference implements Comparable<RelativeBusinessObjectReference> {
-	private ImmutableList<String> segments;
-	private ImmutableList<String> lcSegments; // Lowercase segments. Used for comparison.
-
+public class RelativeBusinessObjectReference extends BusinessObjectReference {
 	/**
 	 * Creates an instance from an array of segments. Segments are case insensitive. Throws an exception is optional or if the segments
 	 * array is null or empty.
 	 * @param segments is an array of segments that makes up the reference.
 	 */
 	public RelativeBusinessObjectReference(final String... segments) {
-		if (segments == null || segments.length < 1) {
-			throw new IllegalArgumentException("segments must contain at least one segment");
-		}
-
-		// Check that all segments are non-null
-		for (final String seg : segments) {
-			if (seg == null) {
-				throw new IllegalArgumentException("segment is null");
-			}
-		}
-
-		// Store segments and lower case segments in separate lists. The lowercase segments are used for comparison while the other list is used to preserve
-		// case for serialization.
-		this.segments = ImmutableList.copyOf(segments);
-		this.lcSegments = this.segments.stream().map(s -> s.toLowerCase(Locale.ROOT))
-				.collect(ImmutableList.toImmutableList());
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((lcSegments == null) ? 0 : lcSegments.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj) {
-			return true;
-		}
-		if (obj == null) {
-			return false;
-		}
-		if (getClass() != obj.getClass()) {
-			return false;
-		}
-
-		final RelativeBusinessObjectReference other = (RelativeBusinessObjectReference) obj;
-		return lcSegments.equals(other.lcSegments);
-	}
-
-	@Override
-	public String toString() {
-		return Arrays.toString(segments.toArray());
-	}
-
-	@Override
-	public int compareTo(final RelativeBusinessObjectReference o) {
-		for (int i = 0; i < lcSegments.size(); i++) {
-			// Check that the reference with which this is being compared to has at least the current number of segments.
-			if (o.lcSegments.size() <= i) {
-				return -1;
-			}
-
-			final int result = lcSegments.get(i).compareTo(o.lcSegments.get(i));
-			if(result != 0) {
-				return result;
-			}
-		}
-
-		return 0;
-	}
-
-	/**
-	 * Returns an immutable list containing the segments.
-	 * @return an immutable list of segments.
-	 */
-	public List<String> getSegments() {
-		return segments;
+		super(segments);
 	}
 
 	/**
 	 * Converts the instance to an instance of the serialized diagram model type.
 	 * @return the converted object.
 	 */
-	public org.osate.ge.diagram.RelativeBusinessObjectReference toMetamodel() {
+	public final org.osate.ge.diagram.RelativeBusinessObjectReference toMetamodel() {
 		final org.osate.ge.diagram.RelativeBusinessObjectReference newValue = new org.osate.ge.diagram.RelativeBusinessObjectReference();
-		for(final String seg : segments) {
+		for (final String seg : getSegments()) {
 			newValue.getSeg().add(seg);
 		}
 		return newValue;
